move file walking/listing logic into the ./files package
