+++ files/init.go
@@ -1,0 +1,9 @@
+package files
+
+import (
+	"log"
+)
+
+func init() {
+	log.SetFlags(log.Flags() | log.Lshortfile)
+}
