+++ storage/disk_fetcher.go
@@ -2,20 +2,15 @@ package storage
 
 import (
 	"bytes"
-	"errors"
 	"io"
 	"log"
-	"os"
 	"path/filepath"
-	"regexp"
+
+	"github.com/arschles/goprox/files"
 )
 
 const (
 	srcDir = "src"
-)
-
-var (
-	errNilFileInfo = errors.New("nil file info")
 )
 
 // DiskFetcher is a Fetcher that reads from the gopath given. it ignores all versions and
@@ -30,7 +25,7 @@ func (d DiskFetcher) GetContents(pkgName string, version string) (io.Reader, err
 	log.Printf("package %s", pkgName)
 	goPathSrcDir := filepath.Join(d.Gopath, srcDir)
 	repoPrefix := filepath.Join(goPathSrcDir, pkgName)
-	files, err := getFiles(repoPrefix, d.Excludes...)
+	files, err := files.List(repoPrefix, d.Excludes...)
 	if err != nil {
 		return nil, err
 	}
@@ -40,43 +35,4 @@ func (d DiskFetcher) GetContents(pkgName string, version string) (io.Reader, err
 		return nil, err
 	}
 	return buf, err
-}
-
-// get a list of relative paths of all files under dir. call filepath.Join(dir, file) for each
-// returned file to get the absolute path
-func getFiles(dir string, excludes ...string) ([]string, error) {
-	files := []string{}
-	if err := filepath.Walk(dir, getWalkFunc(dir, &files, excludes...)); err != nil {
-		return nil, err
-	}
-	return files, nil
-}
-
-func getWalkFunc(baseDir string, files *[]string, excludes ...string) filepath.WalkFunc {
-	return func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			// TODO: handle this case!
-			return nil
-		}
-		if info == nil {
-			// TODO: handle this case!
-			return errNilFileInfo
-		}
-		if info.IsDir() {
-			return nil
-		}
-		rel, err := filepath.Rel(baseDir, path)
-		if err != nil {
-			return err
-		}
-		for _, exclude := range excludes {
-			matched, err := regexp.Match(exclude, []byte(rel))
-			if err == nil && matched {
-				log.Printf("excluding file %s", rel)
-				return nil
-			}
-		}
-		*files = append(*files, rel)
-		return nil
-	}
 }
